3d-tracking/requirements.txt
torch==0.4.1
torchvision==0.2.1
tqdm==4.25.0
cffi==1.11.5
h5py==2.8.0rc1
Cython==0.28.2
munkres==1.0.12
opencv-python==3.4.0.12
numpy==1.14.3
numba==0.39.0
matplotlib==2.1.2
easydict==1.7
progressbar2==3.38.0
filterpy==1.4.4
scikit_learn==0.19.2
ruamel-yaml==0.15.81
PyYAML==3.13
shapely==1.6.4
seaborn==0.9.0
jupyter==1.0.0
utm==0.4.2
tensorboardX==1.4
joblib==0.13.0
panda==0.3.1
requests==2.20.1
xlrd==1.1.0
